Remove AxisPtr optimization
The result is 15% slower Fills for one-dimensional histograms, and
up to 42% for templated Fills. For higher dimensions, the difference
is small or sometimes the direct approach is now faster. Overall,
the one-dimensional case is still fast so prefer the simpler version.

In spirit, this reverts commit 68cb161834 ("Cache direct pointers to
axis objects") and commit c8b87ff4d7 ("Use reinterpret_cast instead
of std::get").